Move the TransformedComponent offset hackery to dirty().
This is the method that gets called by the super constructor.

I might introduce some sort of "we are totally dirty and we've told our parent
about it" flag, because otherwise when you create a list view, every single
sub-element creation triggers one or more traversals all the way up to the
root, merging dirty rectangles and doing a lot of needless math.

This would also eliminate the need for this _offset hackery because elements
would start out dirty and we wouldn't have to worry about getting asked to
fiddle with dirty regions during UI creation.