Fix Scroller/Panner preferred size.
Elements should never prefer their hint sizes. Either they prefer 0x0 or they
prefer some size based on their contents (like a text label prefers to fit its
text, a group prefers to fit all of its children's preferred sizes).

With this fix, the graph editors would now properly size themselves to fit
their contents (without exceeding the hint/max size which in this case is
configured to be 90% of the screen size). But we want the graph editors to be
the full screen size regardless of the size of their contents, so we supply
their hint/max sizes also as their min sizes.