Improve host event handling.
Various assumptions were being made about event coordinates vis a vis root
origins and the HTML elements that hosted everything. Things are still
complicated, but the hosts are now in charge of converting browser event
coordinates into root coordinates and can do the right thing based on what DOM
element they're listening to and where the roots are rendered on the page.

Some of this complexity stems from the annoying fact that we can't listen for
mouseup on HTML elements, we have to listen for those on the whole document,
because if you mouseup outside the bounds of the element you're listening on,
you just don't hear about it. Thus MouseEvent.offsetX/Y is in a different
coordinate system for mouseup events. Of course browsers "fixed" this for touch
events, you get touchend even if the touch ends outside your element bounds,
but we're stuck with unfortunate decisions made in the last millenium when it
comes to mice.

In fact, now that I complain about this, I think I'm just going to redo all
event handling to use events from document. We already have to do the math, so
we might as well just simplify everything and use one set of listeners.