Maintain per-Root hover sets.
Otherwise things go haywire when we have more than one root on screen.

This also changes applyToContaining and applyToIntersecting to use expanded
bounds in element (which seems correct since we use that for all other bounds
tests) and to avoid descending into children unless our own bounds (and
visibility) checks passed.

This assumes that children are entirely contained by their parents, but I
believe that we want to assume that.