Rewrite publish_snapshot in JavaScript.
The Jenkins builds run in Docker, which is running on Alpine Linux (so that the
Docker images are like 100MB instead of 900MB) aaaand Alpine doesn't ship with
bash. But we have Node because that's needed to do the actual build. So in this
case, two wrongs do make a right!

Not sure if we have /usr/bin/env but hey, we'll find out!