New approach to snapshot builds.
NPM apparently hates continuous integration because it doesn't have a first
class concept of a snapshot build and if you try to fake one, assumptions are
violated everywhere and you quickly hate your life.

So instead we just bump a version number for every build and then deal with the
fallout of having zillions of fucking copies of our 250k package all over the
place because that's at least something under our control. Yay.