Allow a 'select' property to specify options as a Map<string, any>.
Ideally, IMO, this would be a Map<any, string>. It maps unique
selectable values to possibly-duplicated strings. But I guess: whatever,
because a user can't tell the difference between two of the same string
anyway.

Perhaps, rather than complicating things, we should just make two
different things, maybe one is called 'select' and just uses a string[]
and 'option' could be this new thing. Right now I remap string[]s:
  new Map<string,string>(options.map(opt => [opt, opt]))

Someone may want to review that I'm doing the right thing when a value
is selected. It looks up the value then in the map... but at that later
time. In theory, if the map changed, the data would all re-resolve, so
we *know* that the map is the same at that later time. Right?