Unified keys + ModelProvider into ElementsModel.
This simplifies a lot of plumbing and in the case of dataModel() when used to
create menu shortcuts yields very nice DRY.

A lot of stringly typed code had to change, but I'm pretty sure I got
everything. If something is weird, let me know. I think this simplification is
going to pay sufficient dividends to be worth the pain of conversion today.