The key of ElementsModel need not be constrained.
The creator of the model can make sure that when a key from keys is passed to
resolve that the right thing happens (the model making helper functions do
constrain K because they stick the models in a map). Then at the point where
the model is part of an ElementConfig it can also be constrained (ListConfig
for example says it needs an ElementModel that conforms to ModelKey because it
too sticks keys in a map).

But if you are in control of the horizontal and the vertical, then you can do
what you like.