Allow dynamic choice in config of list element.
You can supply a function which is passed the element model and the config for
the element UI can be based on some dynamic component of the element model.
It's not fully dynamic in that you can't change the config of the list element
based on some reactive value of the element model, but you can at least have
elements of different types and choose the UI based on the type at the time the
element comes into existence.