Put the screen bounds in UINodeContext.
Then use that to size the root based on the whole bounds. This moves the "fill
most of the screen" logic into one place, and does the right thing with
non-full-screen bounds.

This also swaps out origin + size with a full rootBounds in the UINodeConfig.
So either you say where and how big your root is, or we center it in the screen
bounds provided to the node context. This avoids weird cases where someone
could, say, specify the origin and not the size and there wouldn't be any
sensible way for us to decide what size to make the view.