Make sizeValue a proper Value.
The implementations using fold were very problematic. I'm kind of tempted to
take fold away because the fact that it doesn't update when it has no listeners
makes it very likely to violate someone's expectations.

Maybe it should just return a Subject which better conveys the expectations
that you have to be listening.