Improvements to action system and key bindings.
A model can now contain bare functions as actions, or "commands" which are a
function and an enabled state. Elements that bind to actions (buttons and the
things that extend button, text fields, etc.) are automatically bound to the
enabled state of their target command (if it exists).

Additionally, every Root now has a Keymap which it uses to dispatch key events
that are not handled by the focused element. Elements can push key bindings
onto this keymap to have actions invoked based on key presses. The menu system
now uses this for keyboard shortcuts.

I was tempted to push this further and have elements push key bindings when
they receive focus and pop them when the lose focus, and eliminate the manual
handling of key events entirely. However, the only element that even accepts
focus at the moment is the text editing field, which is currently totally
usurped by the HTML5 overlay so that things work on mobile.

I left the door open to going completely to key-binding based keyboard
interaction, including for the focus widget, and if we eventually support
keyboard navigation for focus and focusing things like buttons and checkboxes
and toggling/triggering those with the keyboard, then we can revisit.