Add ElementConfigMaker & use for dynamic lists & tabs.
This allows element configuration to vary based on the element key or even
information in its model (though only at the time the element is created, one
cannot bind the element config to some reactive model element).

This also allows us to pass the key down into draggable dynamic components
without burdening the user of said component with sticking it in the model and
telling us about it.