Fix auto-activate.
Previously we only tracked open children, but we want to track active children,
but still only auto-activate (a new child) if the active child is open.

Also fixed clicking a menu to close it again. This is "right" for the menu bar
and "differs from macOS behavior" for submenus (they remain open even if you
click them), but it seems like not a big deal. If it bugs anyone I can add the
extra wiring to make it noop on submenu clicks.