Use a floating root for drag and drop.
This allows elements to be dragged outside their host root. This is nice even
if the element ultimately gets dropped on its host root, it's weird to have
them disappear inside the app or render behind other elements that happen to
render after the drag source.

This also factors the cursor stuff out of text and uses it for the drop cursor,
which makes for more straightforward styling and will make it easier to have
fancier drop cursors in the future if we so desire.