Fix active menu child handling.
If a menu item is unhovered and does not have its menu open, it needs to tell
its host that it's not the active child any longer. If it does have its menu
open, it remains the active child even when unhovered because we don't want to
pop down its menu until the user either pops up someone else's menu or
dismisses the popped up menu via other means.