Style resolution improvements.
Previously I was listening for state changes and the manually rebinding
observers to the current set of resolved styles. This resulted in recreation of
things like backgrounds and borders and whatnot every time the state changed,
even if the style for the new state mapped to the exact same style config.

Now we map the state to the style config for that style, and use an equality
checked that will suppress updates unless the style config is actually
different, then we map that value to the resolved style and observe that. So
now nothing gets rebuilt (and the elements do not get spuriously invalidated)
unless something *actually* changes.