Revamp element bounds handling.
There are three distinct bounds that we care about:

* layout bounds: controls how an element is laid out in a container

* hit bounds: controls whether an input event "hits" an element and interacts
  with it

* render bounds: tells the rendering system what area is painted on by an
  element when it is rendered

These are all tracked separately now and computed in expandBounds(). This
bounds recomputation is updated automatically when an element is validated, but
elements that change their bounds outside the invalidation/validation cycle can
manually call recomputeBounds().

This also adds Elements.applyToChildren and uses that to simplify the
revalidation process (after relayout is called, all children are validated
automatically) and disposal (all children are disposed when an element is
disposed). This reduces the number of places an element has to hook in to tell
the system about its children.