Added support for overlapping gestures.
A button press or touch can now start multiple pointer interactions, and as
those pointer interactions process move events, they can opt to take over the
interaction. This allows a press to start on a button, say, but then turn into
a scroll if the user drags far enough to cause the scroller to take over the
interaction.

This also adds support for adding gesture handlers to roots, so that gestures
in addition to the default element gestures can be processed.