Resolve styles on demand in leaf-class constructor.
This avoids the earlier problem that we needed to have an instance of the
Element before we could properly know what it's style scope was. This was
mostly glossed over because usually a Control subclass would define a style
scope and only its children would use it, so the Control subclass would not
actually have its styles resolved using its custom scope even though its
children would.

Now an Element subclass specifically requests its styles to be resolved inside
its own constructor, at which point we can know what its custom scope id is.
Control subclasses properly get their styles resolved with their custom scope
id and the hackery to make TerminalStyleScope work can go away.

This removes EdgeViewStyleScope because it appeared not to be used. It was
defined on EdgeView which never used it because of the "scope definer doesn't
actually get their defined scope" problem, and it has no children.

As a minor fringe benefit, we no longer need to rewrite element configs to
inject the resolved styles, and elements that don't actually use styles (which
is a lot of them) don't even try to do the style resolution process, which can
be quite involved.