Make model resolution more explicit.
Silent failures are not OK. `foo is undefined` deep in the bowels of the UI
toolkit errors are not OK.

If your UI config is missing a required parameter, we now give you a useful
error message that tells you exactly what you did wrong. That means that
elements have to be more explicit about whether config elements are optional or
not when resolving them.