Support null as a codec type explicitly.
Annoying that JavaScript has two types of null, but preserving it properly will
avoid unexpected conversions as things go over the wire.

This also revamps codec registration to better reflect the fact that encoded
data may be stored persistently (we already do so in Firebase for example). So
built-in codecs are assigned ids explicitly and registered using their explicit
ids and registerCustomCodec is modified to require a manually assigned id and
its comments reflect that the caller must take care to choose an id that will
never change.

IDs 0 through 63 are reserved for the platform, we currently use 0-10, so
hopefully that's sufficiently forward looking. I can imagine adding support for
typed arrays (specifically Float32Array) because we use that for our vector and
matrix types, but that will only require one new id.