Use client coords rather than offset coords for mouse position.
And subtract the canvas position, as we do with touches.  I was driving
myself crazy trying to figure out why there was a discontinuity in camera
movement; turns out I was getting coordinates relative to the canvas until
I pressed the mouse button, at which point the pointer was grabbed and I
started getting coordinates relative to the root element.