Clear elements from hover & focus when disposed.
If a list dynamically deletes an element, the Root needs to know about that so
that it can remove it from the focus and from its hovered element set.

This is bad both because hanging onto a disposed element is a (probably brief)
memory leak, but also that changing its hovered state will trigger a state
recomputation for the deleted element which, when it comes from a list, can
result in the element going looking in a reactive map for its state but that
state no longer exists which is why the element was deleted in the first place.