Support wrapping text in label.
This isn't exactly how I want to do it because I'd like to have Label only
support a single line of text and have a more general purpose RichLabel that
supports styled spans, icons and wrapping all in one super powerful package.

But I'm tired of chat messages not getting wrapped and the RichLabel is going
to take more time than I want to spend right now, so this is a stopgap.