More invalidation edge cases.
Normally when an element is either invalid or invisible we ignore invalidation
requests. If the style or text changes on an invisible element, we don't want
to do a bunch of pointless computation. But when an element becomes invisible
or visible, it needs to trigger invalidation up the chain.

I had previously fixed triggering invalidation when an element became
invisible, but I didn't realize there was also a problem when invisible
elements became visible: they would be invalid at the time, so the needed
parent invalidation to cause them to become valid would never happen.