Manually validate cursor after changing its bounds.
By the time recomputeBounds is called, our children have been validated, so
it's too late to be fooling around changing bounds. But we have to wait for our
Label child to be revalidated so that we can recompute the correct position of
the cursor. So we just have to validate the cursor again after moving it to its
new position so that it can do all the right bounds fiddling and dirty itself
at its new position before we return from validate and move on to render.