Added Styler component.
This allows one to dynamically set the style state of a UI subtree. It requires
a custom scopeId, which is a little bit cumbersome.

It would be nice to just be able to bind a single style configuration to a
reactive value, but there's no great way to do that that doesn't either make
everything super loosey goosey, or require that 99.99% of style configurations
are pointlessly wrapped in Value.constant().