An improvement to ValueBuffer..
Why test if we've been provided with an updater every time when we know
it prior to the constructor? Use a simple default updater and remove
updater checks in the ValueBuffer.

This could also be done with 2 different subclasses, one that does the
simple update and one that uses the updater, but that felt like it was
adding a lot of code weirdness.

This could also be done without the static shared defaultUpdater, just
making one every time... unsure on what's better, there.