Always include at least one character when hard wrapping.
Otherwise we can get degenerate cases where we're provided with less than a
single character's width in which to wrap our text and we infinite loop because
we can't add at least a one character hard wrapped line in the allotted space.
At that point, we have no choice but to spill outside the bounds.