Some file data store robusting.
Write new data to file.new then rename file.new to file, to avoid problems if
we die in the middle of writing a file.

If we fail to read a file just proceed with empty data rather than leaving that
object in a broken state. This is just a dev data store, so I think that's more
useful. You're not going to try to recover the corrupted file.