Error handling improvements.
Because decoding an OBJ response requires reading the data right from the
network into the DObject, we end up dispatching notifications during that
process. We now catch any errors that occur during dispatch and just log them
so that it does not bork the handling of the OBJ message.

Also when dispatching notifications due to single property change, we catch the
error earlier and log it with a bit more useful information instead of letting
it percolate all the way up to the channel system where it would get logged
with internal channel details that were less helpful.

And when dispatching channel messages we report dispatch errors separately from
decoding errors. The former were erroneously being reported as the latter, and
when reporting the latter we can actually include the message that failed to
dispatch.