Bounds computation adjustments in preparation for octree raycasting.
Raycasting in Three.js first checks the bounding sphere (computing it if
necessary), then the bounding box (if present).  Frustum culling uses the
bounding sphere only (again, computing it if necessary).  We're mostly
interested in the bounding box, so compute that for all meshes when we load
models and use the box to compute the sphere.  Then we generate the merged
box for the model, and use the model boxes to compute merged mesh bounds.