Consolidated resource loading API.
A ResourceLoader is now used everywhere to load any sort of resource
(configuration or media). In the case of GLTFs loaded for Three this is only
partially true, the ResourceLoader provides the URL and then Three's loader is
used.

The API supports hot reloading (with a few limitations which we can iron out
later) though only on the server when resources are loaded via the filesystem
are resources currently hot reloaded.

We could eventually wire up a simple hot reloading notification (on dev
systems) whereby the server watches files and tells the client to reload things
when they change. Or if we make an Electron client, then we can just load
resources directly from the file system and use file watching to hot reload
there as well.