Service for having server watch resources for client.
Used on dev, allows resource hot reloading (on local file change) while still
running client in web browser.

This also changes the way we handle watches. We can't just watch files because
if an editor replaces a file by writing new contents to a temp file and
renaming it over the old file, Node immediately loses the plot and no longer
watches that file. So we watch the direectory that contains the file instead
and then react to changes to our target file name.