Don't try to load image at path "".
An empty path in intended to mean "don't show anything". This would presumably
also be used in conjunction with hiding the image element, but we don't want to
not load the image for an image element until it is first visible. One usually
does want preloading behavior such that an image element can be made visible at
any time and its image is shown immediately, not after some loading delay.