Have UI prevent default on mouse events as well.
It already does this on touch and keyboard events.

Also have the stage pointer listener check whether default has been prevented
so that it cooperates more nicely with UI mouse/touch listeners. It already
sets prevent default which would stop the UI code from handling an event that
the stage pointer already handled, but the vice versa was not also handled.