Reinstate pos on release.
I felt it was unnecessary because it's always the same as the last move event,
but maybe you don't care about move events and just want to handle release, so
I'm putting it back.

Scroller also uses it to add the final position a second time, but with the
release timestamp rather than the last move timestamp, which is meaningful.