Add support for text "spill".
When rendering emojis, text has a habit of going *way* outside the bounds it
claims it will be rendering into. So we hackily allow one to supply a "spill
region" which we assume text may spill into and thus expand our render bounds
accordingly. This just dirties a larger area when the text is dirtied, it
doesn't end up taking more space in the UI for the text.