Change exclusivity handling.
We might need to adjudicate between two conflicting drag interactions, but
still allow a tap interaction to overlap with whichever drag interaction takes
priority. So a single "exclusive" flag isn't powerful enough.

Now one assigns an exclusivity "group" and only one interaction in a given
group will be allowed to proceed, but non-grouped or interactions in a
different group are allowed to coexist. This also adds priority so that one
does not have to rely on the order that interaction providers are registered.

Claiming an interaction after the pointer has moved is still all or nothing,
but maybe we'll want that to be more flexible as well? So far that hasn't come
up, but if it does we can perhaps extend this approach to that as well.

I also want to add support for simple multi-touch interactions and coordinating
between say a drag and a pinch-zoom, but that can wait until later.