Reject invalidation attempt during validation.
When a parent is laying out its children, it needs them to invalidate
themselves if appropriate. But between the time that it starts validating those
children and the time it marks itself as valid, it is illegal for the children
to request to be invalidated again.

We could perhaps automatically defer the invalidation to the next frame, but
let's see if/when we ever run into this issue again before papering over it.