Hint the whole width to each column.
The other columns could be zero width, or more likely they take up less than
1/n of the available width. We could allocate a minimum width to each column,
but the layout system is meant to handle changes between estimated layout width
and actual layout width for exactly these reasons.

Unfortunately wrapped text labels are not currently doing the right thing in
this circumstance, but it's not too bad and I'll fix that later.